Share ES exception test between clean-up commands
